Almost finished question generator, modified start function, made and finished play again function
Added onto the question generator function by making it so the other questions work, added images, added a working play again function, modified the start function so that the buttons and image appear from the start function this is because I now made the buttons outside of any functions. Modified question generator function by making it so options are now made outside of any function and  made it so questions appear in a random order.